refactor(issuer2): credential offer models
