+++ waltid-services/waltid-issuer-api2/src/main/kotlin/id/walt/issuer2/controller/openapi/Issuer2ManagementRoutesDocs.kt
@@ -1,7 +1,7 @@
 package id.walt.issuer2.controller.openapi
 
-import id.walt.issuer2.controller.dto.CredentialOfferCreateRequest
-import id.walt.issuer2.controller.dto.CredentialOfferCreateResponse
+import id.walt.issuer2.models.CredentialOfferCreateRequest
+import id.walt.issuer2.models.CredentialOfferCreateResponse
 import id.walt.issuer2.domain.CredentialProfile
 import id.walt.issuer2.domain.IssuanceSession
 import io.github.smiley4.ktoropenapi.config.RouteConfig
