+++ waltid-services/waltid-issuer-api2/src/main/kotlin/id/walt/issuer2/models/CredentialOfferModels.kt
@@ -1,7 +1,7 @@
-package id.walt.issuer2.controller.dto
+package id.walt.issuer2.models
 
-import id.walt.mdoc.dataelement.json.JsonObjectToCborMappingConfig
 import id.walt.issuer2.notifications.IssuanceNotifications
+import id.walt.mdoc.dataelement.json.JsonObjectToCborMappingConfig
 import id.walt.openid4vci.offers.AuthenticationMethod
 import id.walt.openid4vci.offers.CredentialOfferValueMode
 import id.walt.openid4vci.offers.IssuerStateMode
@@ -11,6 +11,8 @@ import kotlinx.serialization.Serializable
 import kotlinx.serialization.json.JsonObject
 import kotlin.time.Duration.Companion.minutes
 
+// The management API and offer service currently use the same data shape.
+// Keep one shared model instead of duplicating DTO and service variants until those contracts diverge.
 @Serializable
 data class CredentialOfferRuntimeOverrides(
     val issuerDid: String? = null,
