+++ waltid-services/waltid-issuer-api2/src/test/kotlin/id/walt/issuer2/testsupport/Issuer2ApiClient.kt
@@ -1,7 +1,7 @@
 package id.walt.issuer2.testsupport
 
-import id.walt.issuer2.controller.dto.CredentialOfferCreateRequest
-import id.walt.issuer2.controller.dto.CredentialOfferCreateResponse
+import id.walt.issuer2.models.CredentialOfferCreateRequest
+import id.walt.issuer2.models.CredentialOfferCreateResponse
 import id.walt.issuer2.domain.CredentialProfile
 import id.walt.issuer2.domain.IssuanceSession
 import id.walt.openid4vci.offers.AuthenticationMethod
